move GenericBackend to ThreadedBackend
